+++ CHANGELOG.md
@@ -13,6 +13,7 @@ Each release can have sections: "Added", "Changed", "Deprecated", "Removed", "Fi
 # added
 
 - test_cli.py and test_get_token.py to cover core API and CLI functionality
+- test_token_ops to test token exchange and JWT verification
 
 # changed